requirements.txt
Django==5.2.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
web3==7.12.0
celery==5.5.3
redis==5.2.1
django-celery-beat==2.8.1
django-celery-results==2.5.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pillow==11.1.0
